Move allocating NumeralString impls into submodule
